requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
httpx==0.27.2
python-dotenv==1.0.1
websockets==13.1
aiosqlite==0.19.0
PyJWT==2.9.0
bcrypt==4.2.0
